frontend/requirements.txt
streamlit
requests
python-dotenv
streamlit-webrtc 
pydub
ffmpeg-python
pillow
streamlit-webrtc 
numpy
streamlit-audiorec